Feature: Can now set arbitrary macros in generated qf file
Calling ->set_macros({ name => 'value', othername => 'value2' })
on a Qf object (or passing a 'macros' key pointing to a hashref in the
constructor) will set those macros in the generated file.

The names of macros should be the bare name, as the module will add any
surrounding {} necessary for multi-character macro names.

Macros are also now emitted in key-sorted order, for consistency and ease of
testing.